+++ src/main/java/org/mesdag/particlestorm/data/molang/FloatMolangExp.java
@@ -30,7 +30,7 @@ public class FloatMolangExp extends MolangExp {
     @Override
     public float calculate(MolangInstance instance) {
         if (!initialized()) return 0.0F;
-        return variable == null ? constant : (float) variable.get(instance);
+        return variable == null ? constant : variable.get(instance);
     }
 
     @Override
