+++ src/main/java/org/mesdag/particlestorm/data/molang/compiler/function/random/RandomIntegerFunction.java
@@ -8,19 +8,16 @@ import org.mesdag.particlestorm.data.molang.compiler.function.MathFunction;
 import java.util.Random;
 import java.util.concurrent.ThreadLocalRandom;
 
-/**
- * {@link MathFunction} value supplier
- *
- * <p>
- * <b>Contract:</b>
- * <br>
- * Returns a random integer value based on the input values:
- * <ul>
- *     <li>A single input generates a value between 0 and that input (exclusive)</li>
- *     <li>Two inputs generates a random value between the first (inclusive) and second input (inclusive)</li>
- *     <li>Three inputs generates a random value between the first (inclusive) and second input (inclusive), seeded by the third input</li>
- * </ul>
- */
+/// [MathFunction] value supplier
+///
+/// **Contract:**
+///
+/// Returns a random integer value based on the input values:
+///
+///   - A single input generates a value between 0 and that input (exclusive)
+///   - Two inputs generates a random value between the first (inclusive) and second input (inclusive)
+///   - Three inputs generates a random value between the first (inclusive) and second input (inclusive), seeded by the third input
+///
 public final class RandomIntegerFunction extends MathFunction {
     private final MathValue valueA;
     @Nullable
@@ -45,9 +42,9 @@ public final class RandomIntegerFunction extends MathFunction {
     }
 
     @Override
-    public double compute(MolangInstance instance) {
+    public float compute(MolangInstance instance) {
         int result;
-        int valueA = (int)Math.round(this.valueA.get(instance));
+        int valueA = Math.round(this.valueA.get(instance));
         Random random;
 
         if (this.random != null) {
@@ -59,7 +56,7 @@ public final class RandomIntegerFunction extends MathFunction {
         }
 
         if (this.valueB != null) {
-            int valueB = (int)Math.round(this.valueB.get(instance));
+            int valueB = Math.round(this.valueB.get(instance));
             int min = Math.min(valueA, valueB);
             int max = Math.max(valueA, valueB);
