+++ src/main/java/org/mesdag/particlestorm/data/molang/compiler/function/round/LerpFunction.java
@@ -5,14 +5,11 @@ import org.mesdag.particlestorm.api.MolangInstance;
 import org.mesdag.particlestorm.data.molang.compiler.MathValue;
 import org.mesdag.particlestorm.data.molang.compiler.function.MathFunction;
 
-/**
- * {@link MathFunction} value supplier
- *
- * <p>
- * <b>Contract:</b>
- * <br>
- * Returns the first value plus the difference between the first and second input values multiplied by the third input value
- */
+/// [MathFunction] value supplier
+///
+/// **Contract:**
+///
+/// Returns the first value plus the difference between the first and second input values multiplied by the third input value
 public final class LerpFunction extends MathFunction {
     private final MathValue min;
     private final MathValue max;
@@ -32,7 +29,7 @@ public final class LerpFunction extends MathFunction {
     }
 
     @Override
-    public double compute(MolangInstance instance) {
+    public float compute(MolangInstance instance) {
         return Mth.lerp(this.delta.get(instance), this.min.get(instance), this.max.get(instance));
     }
